allows cog to run from folders deeper than the project root
this is helpful if your project root/bin folder is not the same as where your typical PWD is. For example, we'll usually have the project root contain the bin folder but the `artisan` or `./craft` CLIs live inside a `html` directory. In this case you could now run `cog` from inside the `html` dir and it would still find `../bin/cog-*` files.